+++ gulpfile.babel.js
@@ -1,4 +1,4 @@
-import qw from 'js-utils/qw';
+import {qw} from 'js-utils';
 import gulp from 'gulp';
 import lazyReq from 'lazyreq';
 
@@ -10,14 +10,20 @@ const $ = lazyReq(require, {
 	mocha: 'gulp-mocha',
 	istanbul: 'gulp-istanbul',
 	isparta: 'isparta',
+	sourcemaps: 'gulp-sourcemaps',
+	uglify: 'gulp-uglify',
 });
 
 gulp.task('build', () =>
 	gulp.src('./src/**/*.js')
+		.pipe($.sourcemaps.init())
 		.pipe($.newer('./build'))
 		.pipe($.babel({
-			sourceMaps: 'inline',
+			babelrc: false,
+			presets: ['es2015'],
 		}))
+		.pipe($.uglify())
+		.pipe($.sourcemaps.write('./'))
 		.pipe(gulp.dest('./build'))
 );